requirements.txt
fastapi
uvicorn
pandas
python-docx
docx2pdf
